Move cluster related types and function into seperate files
- This is with a mind to ban circular dependencies

- With this change all the types and helper functions can be imported
  from anywhere without circular imports
