+++ src/common/utils/cluster-id-url-parsing.ts
@@ -19,42 +19,32 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+import type { ClusterId } from "../cluster-types";
+
 /**
- * A function that does nothing
+ * Grab the `ClusterId` out of a host that was generated by `getClusterFrameUrl`, or nothing
+ * @param host The host section of a URL
+ * @returns The `ClusterId` part of the host, or `undefined`
  */
-export function noop<T extends any[]>(...args: T): void {
-  return void args;
+export function getClusterIdFromHost(host: string): ClusterId | undefined {
+  // e.g host == "%clusterId.localhost:45345"
+  const subDomains = host.split(":")[0].split(".");
+
+  return subDomains.slice(-2, -1)[0]; // ClusterId or undefined
 }
 
-export * from "./app-version";
-export * from "./autobind";
-export * from "./base64";
-export * from "./camelCase";
-export * from "./cloneJson";
-export * from "./cluster-id-url-parsing";
-export * from "./debouncePromise";
-export * from "./defineGlobal";
-export * from "./delay";
-export * from "./disposer";
-export * from "./downloadFile";
-export * from "./escapeRegExp";
-export * from "./extended-map";
-export * from "./getRandId";
-export * from "./hash-set";
-export * from "./local-kubeconfig";
-export * from "./n-fircate";
-export * from "./openExternal";
-export * from "./paths";
-export * from "./reject-promise";
-export * from "./singleton";
-export * from "./sort-compare";
-export * from "./splitArray";
-export * from "./tar";
-export * from "./toggle-set";
-export * from "./toJS";
-export * from "./type-narrowing";
-export * from "./types";
+/**
+ * Get the OpenLens backend routing host for a given `ClusterId`
+ * @param clusterId The ID to put in front of the current host
+ * @returns a new URL host section
+ */
+export function getClusterFrameUrl(clusterId: ClusterId) {
+  return `//${clusterId}.${location.host}`;
+}
 
-import * as iter from "./iter";
-
-export { iter };
+/**
+ * Get the result of `getClusterIdFromHost` from the current `location.host`
+ */
+export function getHostedClusterId(): ClusterId | undefined {
+  return getClusterIdFromHost(location.host);
+}
