+++ src/common/utils/local-kubeconfig.ts
@@ -19,42 +19,15 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-/**
- * A function that does nothing
- */
-export function noop<T extends any[]>(...args: T): void {
-  return void args;
+import { app, remote } from "electron";
+import path from "path";
+import * as uuid from "uuid";
+import type { ClusterId } from "../cluster-types";
+
+export function storedKubeConfigFolder(): string {
+  return path.resolve((app || remote.app).getPath("userData"), "kubeconfigs");
 }
 
-export * from "./app-version";
-export * from "./autobind";
-export * from "./base64";
-export * from "./camelCase";
-export * from "./cloneJson";
-export * from "./cluster-id-url-parsing";
-export * from "./debouncePromise";
-export * from "./defineGlobal";
-export * from "./delay";
-export * from "./disposer";
-export * from "./downloadFile";
-export * from "./escapeRegExp";
-export * from "./extended-map";
-export * from "./getRandId";
-export * from "./hash-set";
-export * from "./local-kubeconfig";
-export * from "./n-fircate";
-export * from "./openExternal";
-export * from "./paths";
-export * from "./reject-promise";
-export * from "./singleton";
-export * from "./sort-compare";
-export * from "./splitArray";
-export * from "./tar";
-export * from "./toggle-set";
-export * from "./toJS";
-export * from "./type-narrowing";
-export * from "./types";
-
-import * as iter from "./iter";
-
-export { iter };
+export function getCustomKubeConfigPath(clusterId: ClusterId = uuid.v4()): string {
+  return path.resolve(storedKubeConfigFolder(), clusterId);
+}
