+++ src/main/catalog-sources/kubeconfig-sync.ts
@@ -27,18 +27,19 @@ import fs from "fs";
 import path from "path";
 import fse from "fs-extra";
 import type stream from "stream";
-import { Disposer, ExtendedObservableMap, iter, Singleton } from "../../common/utils";
+import { Disposer, ExtendedObservableMap, iter, Singleton, storedKubeConfigFolder } from "../../common/utils";
 import logger from "../logger";
 import type { KubeConfig } from "@kubernetes/client-node";
 import { loadConfigFromString, splitConfig } from "../../common/kube-helpers";
 import { Cluster } from "../cluster";
 import { catalogEntityFromCluster, ClusterManager } from "../cluster-manager";
 import { UserStore } from "../../common/user-store";
-import { ClusterStore, UpdateClusterModel } from "../../common/cluster-store";
+import { ClusterStore } from "../../common/cluster-store";
 import { createHash } from "crypto";
 import { homedir } from "os";
 import globToRegExp from "glob-to-regexp";
 import { inspect } from "util";
+import type { UpdateClusterModel } from "../../common/cluster-types";
 
 const logPrefix = "[KUBECONFIG-SYNC]:";
 
@@ -85,7 +86,7 @@ export class KubeconfigSyncManager extends Singleton {
     )));
 
     // This must be done so that c&p-ed clusters are visible
-    this.startNewSync(ClusterStore.storedKubeConfigFolder);
+    this.startNewSync(storedKubeConfigFolder());
 
     for (const filePath of UserStore.getInstance().syncKubeconfigEntries.keys()) {
       this.startNewSync(filePath);
@@ -216,7 +217,7 @@ export function computeDiff(contents: string, source: RootSource, filePath: stri
 
           const entity = catalogEntityFromCluster(cluster);
 
-          if (!filePath.startsWith(ClusterStore.storedKubeConfigFolder)) {
+          if (!filePath.startsWith(storedKubeConfigFolder())) {
             entity.metadata.labels.file = filePath.replace(homedir(), "~");
           }
           source.set(contextName, [cluster, entity]);
