+++ src/main/cluster-detectors/cluster-id-detector.ts
@@ -21,7 +21,7 @@
 
 import { BaseClusterDetector } from "./base-cluster-detector";
 import { createHash } from "crypto";
-import { ClusterMetadataKey } from "../cluster";
+import { ClusterMetadataKey } from "../../common/cluster-types";
 
 export class ClusterIdDetector extends BaseClusterDetector {
   key = ClusterMetadataKey.CLUSTER_ID;
