+++ src/main/initializers/ipc.ts
@@ -23,7 +23,8 @@ import type { IpcMainInvokeEvent } from "electron";
 import { KubernetesCluster } from "../../common/catalog-entities";
 import { clusterFrameMap } from "../../common/cluster-frames";
 import { clusterActivateHandler, clusterSetFrameIdHandler, clusterVisibilityHandler, clusterRefreshHandler, clusterDisconnectHandler, clusterKubectlApplyAllHandler, clusterKubectlDeleteAllHandler, clusterDeleteHandler } from "../../common/cluster-ipc";
-import { ClusterId, ClusterStore } from "../../common/cluster-store";
+import { ClusterStore } from "../../common/cluster-store";
+import type { ClusterId } from "../../common/cluster-types";
 import { appEventBus } from "../../common/event-bus";
 import { ipcMainHandle } from "../../common/ipc";
 import { catalogEntityRegistry } from "../catalog";
