+++ src/renderer/components/cluster-settings/components/cluster-metrics-setting.tsx
@@ -25,8 +25,9 @@ import { Select, SelectOption } from "../../select/select";
 import { Icon } from "../../icon/icon";
 import { Button } from "../../button/button";
 import { SubTitle } from "../../layout/sub-title";
-import { Cluster, ClusterMetricsResourceType } from "../../../../main/cluster";
+import type { Cluster } from "../../../../main/cluster";
 import { observable, reaction, makeObservable } from "mobx";
+import { ClusterMetricsResourceType } from "../../../../common/cluster-types";
 
 interface Props {
   cluster: Cluster;
